refactor: PieceSymbol 객체 dto 패키지로 이동
