+++ src/main/java/chess/dto/PieceSymbol.java
@@ -1,4 +1,13 @@
-package chess.domain.piece;
+package chess.dto;
+
+import chess.domain.piece.Bishop;
+import chess.domain.piece.Color;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
 
 import java.util.Arrays;
