+++ src/test/java/chess/domain/piece/PieceSymbolTest.java
@@ -1,5 +1,6 @@
 package chess.domain.piece;
 
+import chess.dto.PieceSymbol;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
